# swagger.yaml
openapi: 3.0.0
paths:
...
  /templates:
    get:
      operationId: TemplatesController_getContactsV1
      parameters:
        - name: search
          required: false
          in: query
          schema:
            type: string
        - name: nextPageToken
          required: false
          in: query
          schema:
            default: null
            type: string
        - name: limit
          required: false
          in: query
          schema:
            default: 50
            example: 50
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplatesDto"
      security: &a12
        - bearer: []
      summary: Get list of templates
      tags: &a13
        - Templates
    post:
      operationId: TemplatesController_createTemplateV1
      parameters: []
      requestBody:
        required: true
        description: New Template DTO
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateCreateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDto"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDto"
      security: *a12
      summary: Create new template
      tags: *a13
  /templates/{id}:
    get:
      operationId: TemplatesController_getTemplateV1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDto"
      security: *a12
      summary: Get template by ID
      tags: *a13
    patch:
      operationId: TemplatesController_updateTemplateV1
      parameters:
        - name: id
          required: true
          in: path
          description: Template ID
          schema:
            type: number
      requestBody:
        required: true
        description: Template partial data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdateDto"
      responses:
        "200":
          description: Template updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDto"
      security: *a12
      summary: Update a template
      tags: *a13
    delete:
      operationId: TemplatesController_deleteTemplateV1
      parameters:
        - name: id
          required: true
          in: path
          description: Template ID
          schema:
            type: number
      responses:
        "204":
          description: Template deleted successfully
        "404":
          description: Template requested not found
      security: *a12
      summary: Delete a template
      tags: *a13
...
info:
  title: OnPage API
  description: >-
    OpenAPI definition for managing dispatching, pages, contacts,
    administrators, templates, settings, and webhooks.


    [📥 Download Swagger YAML](./swagger.yaml)
  version: "1.0"
  contact: {}
...
servers:
  - url: http://127.0.0.1:3003
    description: Development Server
  - url: https://public-api.onsetmobile.com
    description: Sandbox Server
  - url: https://public-api.onpage.com
    description: Production Server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Metadata:
      type: object
      properties:
        nextPageToken:
          type: string
      required:
        - nextPageToken
...
    TemplateDto:
      type: object
      properties:
        id:
          type: number
          description: Template ID
        name:
          type: string
          description: Template name
        subject:
          type: string
          description: Template subject
        body:
          type: string
          description: Template body
        predefinedReplies:
          description: Predefined replies
          type: array
          items:
            type: string
        syncToDevice:
          type: boolean
          description: Sync to device flag
          default: false
      required:
        - id
        - name
        - subject
    TemplatesDto:
      type: object
      properties:
        templates:
          description: List of templates
          type: array
          items:
            $ref: "#/components/schemas/TemplateDto"
        metadata:
          description: Response metadata
          allOf:
            - $ref: "#/components/schemas/Metadata"
      required:
        - templates
        - metadata
    TemplateCreateDto:
      type: object
      properties:
        name:
          type: string
          description: Template name
        subject:
          type: string
          description: Template subject
        body:
          type: string
          description: Template body
        predefinedReplies:
          description: Predefined replies
          type: array
          items:
            type: string
        syncToDevice:
          type: boolean
          description: Sync to device flag
          default: false
      required:
        - name
        - subject
    TemplateUpdateDto:
      type: object
      properties:
        name:
          type: string
          description: Template name
          example: Template name
        subject:
          type: string
          description: Template subject
        body:
          type: string
          description: Template body
        predefinedReplies:
          description: Predefined replies
          type: array
          items:
            type: string
        syncToDevice:
          type: boolean
          description: Sync to device flag
...